Moved get_complementary_ops function outside of the if elif statement that it was in.
Changed return statements to yield statements in get_complementary_ops to allow for multiple complementary vis' to be created in the add_filter function found in filter.py.

To allow that change to take effect I also modified the code that creates the complementary vis' by putting it in a for loop so that a new vis is created for each filter returned by the get_complementary_ops function.